refactor(pr06): fix file/class casing for Token + LogHelper (G5e)
- tokenDnd35e.mts -> TokenDnd35e.mts (Windows case-only, two-step git mv)

- class tokenDnd35e -> TokenDnd35e

- logHelper.mts -> LogHelper.mts (Windows case-only, two-step git mv)

Updates sibling index.mts files and all importers.
