+++ src/canvas/index.mts
@@ -1,6 +1,6 @@
 import type { CanvasDnd35e } from './CanvasDnd35e.mjs';
 import { RegionDnd35e } from './RegionDnd35e.mjs';
-import { tokenDnd35e } from './token/tokenDnd35e.mjs';
+import { TokenDnd35e } from './token/TokenDnd35e.mjs';
 
 export type {
   CanvasDnd35e,
@@ -8,5 +8,5 @@ export type {
 
 export {
   RegionDnd35e,
-  tokenDnd35e,
+  TokenDnd35e,
 };
