+++ src/canvas/token/TokenDnd35e.mts
@@ -1,8 +1,8 @@
 import type TokenLayer from '@client/canvas/layers/tokens.mjs';
 import type { TokenDocumentDnd35e } from '@scene/token-document/TokenDocumentDnd35e.mjs';
 
-class tokenDnd35e<TDocument extends TokenDocumentDnd35e = TokenDocumentDnd35e> extends fc.placeables.Token<TDocument> {
+class TokenDnd35e<TDocument extends TokenDocumentDnd35e = TokenDocumentDnd35e> extends fc.placeables.Token<TDocument> {
   declare readonly layer: TokenLayer<this>;
 }
 
-export { tokenDnd35e };
+export { TokenDnd35e };
