+++ src/helpers/index.mts
@@ -12,7 +12,7 @@ import {
 import { buildDocumentDataMap, resolveFormulaField } from './formulae/index.mjs';
 import type { HasSystem } from './HasSystem.mjs';
 import { preLocalizeConfig, registerConfigPreLocalization } from './localization/preLocalizeConfig.mjs';
-import { LogHelper } from './logHelper.mjs';
+import { LogHelper } from './LogHelper.mjs';
 import { parseNumericChangeValue, resolveActiveEffectChanges, STACK_RESULT_APPLIED, STACK_RESULT_IGNORED } from './stacking.mjs';
 import { createTag } from './stringHelpers.mjs';
